+++ src/api/client/auth.ts
@@ -6,7 +6,7 @@ import type {
   SignUpFormValues,
   SignUpResult,
 } from "@/types";
-import axiosInstance from "@/auth/fetcher.client";
+import axiosInstance from "@/lib/auth/fetcher.client";
 
 const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL;
