+++ src/api/client/og.ts
@@ -1,4 +1,4 @@
-import axiosInstance from "@/auth/fetcher.client";
+import axiosInstance from "@/lib/auth/fetcher.client";
 
 export async function getOgData(url: string) {
   const response = await axiosInstance.get(
