+++ src/api/client/posts.ts
@@ -1,4 +1,4 @@
-import axiosInstance from "@/auth/fetcher.client";
+import axiosInstance from "@/lib/auth/fetcher.client";
 import { filterThreadPosts } from "@/lib/postUtils";
 import {
   GetPostsParams,
