+++ src/api/server/meetings.ts
@@ -1,4 +1,4 @@
-import { serverFetch } from "@/auth/fetcher.server";
+import { serverFetch } from "@/lib/auth/fetcher.server";
 import type { JoinedMeetingsResponse, MeetingSortBy, SortOrder } from "@/types";
 
 export async function getJoinedMeetingsServer(params: {
