+++ src/app/meetings/[id]/_components/RecommendedMeetingsSection.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import type { RecommendedMeetingsSectionProps } from "@/types";
 import { Sparkles, ArrowUpRight, ArrowRight } from "lucide-react";
-import FallbackImage from "@/components/img/FallbackImage";
+import FallbackImage from "@/components/ui/FallbackImage";
 import { useMeetingRecommendationsQuery } from "@/app/meetings/[id]/_hooks/useMeetingDetail";
 
 import { useDragScroll } from "@/hooks/useDragScroll";
