+++ src/app/ranking/_components/TopRankCard.tsx
@@ -5,7 +5,7 @@ import { Card } from "@/components/ui/Card";
 import { TopRankCardProps } from "@/types";
 
 import { cn } from "@/lib/utils";
-import FallbackImage from "@/components/img/FallbackImage";
+import FallbackImage from "@/components/ui/FallbackImage";
 import { ArrowUpRight } from "lucide-react";
 
 export default function TopRankCard({
