+++ src/app/users/[id]/_components/ProfileSection.tsx
@@ -8,9 +8,9 @@ import { useAuthStore } from "@/store/useAuthStore";
 import ModalBase from "@/components/modal/ModalBase";
 import { Input } from "@/components/ui/Input";
 import { Button } from "@/components/ui/Button";
-import { ImageUploadInput } from "@/components/img/ImageUploadInput";
+import { ImageUploadInput } from "@/components/ui/ImageUploadInput";
 import { Settings2 } from "lucide-react";
-import FallbackImage from "@/components/img/FallbackImage";
+import FallbackImage from "@/components/ui/FallbackImage";
 import { ConfirmModal } from "@/components/modal/ConfirmModal";
 import { useUpdateProfile } from "@/app/users/[id]/_hooks/useUser";
 import { useRouter } from "next/navigation";
