+++ src/app/users/[id]/_components/UserCard.tsx
@@ -5,7 +5,7 @@ import { Users, Calendar, LucideLock, Lock } from "lucide-react";
 import { Card } from "@/components/ui/Card";
 import { UserCardProps } from "@/types";
 import { cn } from "@/lib/utils";
-import FallbackImage from "@/components/img/FallbackImage";
+import FallbackImage from "@/components/ui/FallbackImage";
 import { HeartIcon } from "@/components/ui/HeartIcon";
 
 export function UserCard({
