+++ src/bff/ranking.ts
@@ -1,4 +1,4 @@
-import { serverFetch } from "@/auth/fetcher.server";
+import { serverFetch } from "@/lib/auth/fetcher.server";
 import { CursorResponse } from "@/types";
 import { fetchAllCursor } from "@/lib/fetchAllCursor";
