+++ src/bff/recommend.ts
@@ -1,4 +1,4 @@
-import { serverFetch } from "@/auth/fetcher.server";
+import { serverFetch } from "@/lib/auth/fetcher.server";
 import { fetchAllCursor } from "@/lib/fetchAllCursor";
 import { isSecretMeeting } from "@/lib/meetingSecret";
 import type {
