+++ src/tests/meetings/detail/RecommendedMeetingsSection.test.tsx
@@ -33,7 +33,7 @@ jest.mock("next/link", () => ({
   ),
 }));
 
-jest.mock("@/components/img/FallbackImage", () => ({
+jest.mock("@/components/ui/FallbackImage", () => ({
   __esModule: true,
   default: ({ alt }: { alt: string }) => <img alt={alt} />,
 }));
